cashlink,file-handler: remove unnecessary and broken enforcement of /cashlink in cashlink URL
The removed code attempted to preserve /cashlink in the URL (for what ever reason), but
in fact did not only preserve it, but also enforced it to be added. In file-handler, the
original url path would even be discarded entirely.
It was also not very elegant to begin with. A better place for handling of an invariant
like the url enforcing a specific format, would have been the Cashlink constructor instead
of doing and repeating it on parsing and exporting, but then also omitting it on Cashlinks
that are created via the constructor.

See https://github.com/nimiq/cashlink-generator/pull/1#discussion_r1934545356
and https://github.com/nimiq/cashlink-generator/pull/1#discussion_r1935980329