Keep user private key only temporary in memory, don't import them
Revert private key handling to the previous approach where the key was only temporary held
in memory and not imported, which didn't expose the key and was better suited to using a
separate account for each campaign, without leaving behind a bunch of imported keys. Also,
importing the key did't make all that much sense, if the user still has to provide the full
recovery words / private key on usage anyway, as opposed to just providing the funding address
of a key, which then only asks for the recovery words, if it's not imported yet.

See https://github.com/nimiq/cashlink-generator/pull/1#discussion_r1937968924